+++ test/minifier_test.dart
@@ -25,4 +25,46 @@ void main() => group('Minifier', () {
       expect(await minifier.transformer.transform(script), contains('__construct() { }'));
     });
   });
+
+  group('.processDirectory()', () {
+    var minifier = new Minifier(silent: true);
+    var testDir = getDir('var/test/processDirectory');
+
+    tearDownAll(() {
+      delete(testDir);
+      return minifier.transformer.close();
+    });
+
+    test('should remove the comments and whitespace from the scripts of a directory', () async {
+      var output = joinFile(testDir, ['sample.php']);
+      await minifier.processDirectory(getDir('test/fixtures'), testDir);
+      expect(await output.readAsString(), allOf(
+        contains("<?= 'Hello World!' ?>"),
+        contains('namespace dummy; class Dummy'),
+        contains(r'$className = get_class($this); return $className;'),
+        contains('__construct() { }')
+      ));
+    });
+  });
+
+  group('.processFile()', () {
+    var minifier = new Minifier(silent: true);
+    var testDir = getDir('var/test/processFile');
+
+    tearDownAll(() {
+      delete(testDir);
+      return minifier.transformer.close();
+    });
+
+    test('should remove the comments and whitespace from a file', () async {
+      var output = joinFile(testDir, ['sample.php']);
+      await minifier.processFile(getFile('test/fixtures/sample.php'), output);
+      expect(await output.readAsString(), allOf(
+        contains("<?= 'Hello World!' ?>"),
+        contains('namespace dummy; class Dummy'),
+        contains(r'$className = get_class($this); return $className;'),
+        contains('__construct() { }')
+      ));
+    });
+  });
 });