+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file contains highlights of what changes on each version of the [Grinder-PH
 - Breaking change: removed the `binary` and `mode` properties from the `Minifier` class.
 - Breaking change: the `Minifier.transformer` property is now `final`.
 - Added a factory constructor to the `Transformer` class.
+- Added new unit tests.
 - Changed licensing for the [MIT License](https://opensource.org/licenses/MIT).
 - Updated the package dependencies.